+++ include/nstd/gl/surface/config.h
@@ -1,0 +1,9 @@
+#ifndef NSTD_GL_SURFACE_CONFIG_H_INCLUDED
+#define NSTD_GL_SURFACE_CONFIG_H_INCLUDED
+#include "../../core/def.h"
+#include "../../nstd.h"
+
+/// Represents a surface config.
+typedef NSTDAny NSTDGLSurfaceConfig;
+
+#endif