+++ src/gl/surface/config.rs
@@ -1,0 +1,4 @@
+use wgpu::SurfaceConfiguration;
+
+/// Represents a surface config.
+pub type NSTDGLSurfaceConfig = *mut SurfaceConfiguration;